float

+++ ex2/timeseries.cpp
@@ -35,7 +35,7 @@
 						this->rowLength++;
 
             // Initialize and add <colname, int vector> pairs to result
-            this->result.push_back({colname, std::vector<double> {}});
+            this->result.push_back({colname, std::vector<float> {}});
         }
     }
 
@@ -80,23 +80,23 @@
 	}
 
 	// returns spacific column by index
-  const std::pair<std::string, std::vector<double>> TimeSeries::getColumnByIndex(int index) const{
+  const std::pair<std::string, std::vector<float>> TimeSeries::getColumnByIndex(int index) const{
 		   return this->result.at(index);
   }
 
 	// returns spacific column by name
-	const std::vector<double> TimeSeries::getColumnByName(std::string name) {
+	const std::vector<float> TimeSeries::getColumnByName(std::string name) {
 		for(int i = 0; i < getRowLength();i++) {
 			if(this->result.at(i).first.compare(name) == 0) {
 				return result.at(i).second;
 			}
 		}
-		const std::vector<double> v = {};
+		const std::vector<float> v = {};
 		return v;
 	}
 
 	// returns spacific column by row and column
-  const double TimeSeries::getColumnByIndex(int column, int row) {
+  const float TimeSeries::getColumnByIndex(int column, int row) {
 		   if(row > this->colLength || column > this->rowLength || column < 0 || row < 0) {
 				 std::cout << this->colLength << "\n";
 				 std::cout << this->rowLength << "\n";
